+++ demo/app/home/home-view-model.ts
@@ -1,5 +1,13 @@
 import { Observable } from "tns-core-modules/data/observable";
-import { request, setImageParseMethod, ImageParseMethod } from "@klippa/nativescript-http";
+
+import {
+    request,
+    setImageParseMethod,
+    ImageParseMethod,
+    certificatePinningAdd,
+    certificatePinningClear
+} from "@klippa/nativescript-http";
+
 import { ImageSource } from "@nativescript/core/image-source";
 import * as dialogs from "tns-core-modules/ui/dialogs";
 import { newWebsocketConnection, IWebsocketConnection } from "@klippa/nativescript-http/websocket";
@@ -155,5 +163,37 @@ export class HomeViewModel extends Observable {
         if (this.websocket) {
             this.websocket.close(1000, "Goodbye");
         }
+    }
+
+    pinFake() {
+        certificatePinningClear();
+        // We add 2 fake certificate because Trustkit (iOS) requires a backup cert.
+        certificatePinningAdd("*.placeholder.com", ["AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA=", "ABAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA="]);
+        certificatePinningAdd("**.github.com", ["AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA=", "ABAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA="]);
+        certificatePinningAdd("loripsum.net", ["AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA=", "ABAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA="]);
+        this.contentType = "text";
+        this.contentText = "Fake certificates pinned, try to do a request";
+        this.hasContent = true;
+        this.isLoading = false;
+    }
+
+    pinGood() {
+        certificatePinningClear();
+        certificatePinningAdd("*.placeholder.com", ["CDCU5TkA8n3L8+QM7dyTjfRlxWibigF+1cxMzRhlJV4=", "YLh1dUR9y6Kja30RrAn7JKnbQG/uEtLMkBgFF2Fuihg=", "Vjs8r4z+80wjNcr1YKepWQboSIRi63WsWXhIMN+eWys="]);
+        certificatePinningAdd("**.github.com", ["ORH27mxcLwxnNpR7e0i6pdDPWLXdpeWgr5bEfFVbxW8=", "k2v657xBsOVe1PQRwOsHsw3bsGT2VzIqz5K+59sNQws=", "WoiWRyIOVNa9ihaBciRSC7XHjliYS9VwUGOIud4PB18="]);
+        certificatePinningAdd("loripsum.net", ["7ReOzYJ7YC1mMc2CWTuaMDuzynt8xZ+HDQ6K8o+4okk=", "YLh1dUR9y6Kja30RrAn7JKnbQG/uEtLMkBgFF2Fuihg=", "Vjs8r4z+80wjNcr1YKepWQboSIRi63WsWXhIMN+eWys="]);
+
+        this.contentType = "text";
+        this.contentText = "Good certificates pinned, try to do a request";
+        this.hasContent = true;
+        this.isLoading = false;
+    }
+
+    clearPins() {
+        certificatePinningClear();
+        this.contentType = "text";
+        this.contentText = "Certificate pins cleared, try to do a request";
+        this.hasContent = true;
+        this.isLoading = false;
     }
 }